refactor(test): reuse OpenShell release fixtures
